% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2lines.R
\name{df2lines}
\alias{df2lines}
\title{Dataframe to SpatialLines}
\usage{
df2lines(linedf)
}
\arguments{
\item{linedf}{a data frame with the following columns:
id = generic ids of the lines,
fx = coordinates x of the first point of the line,
fy = coordinates y of the first point of the line,
tx = coordinates x of the second point of the line,
tx = coordinates y of the second point of the line.}
}
\value{
SpatialLinesDataFrame
}
\description{
Converts a dataframe of IDs and coordinates into a SpatialLines object of the package "sp".
}
\details{
df2lines
}
\author{
Duccio Aiazzi
}
\keyword{internal}
