% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_shp.R
\name{import_shp}
\alias{import_shp}
\title{Dsn and layer name from .shp filepath}
\usage{
import_shp(path)
}
\arguments{
\item{path}{The full path of a .shp file}
}
\value{
List with two elements: dsn and layer name

List with two character components: dsn and layer
}
\description{
Creates a valid dsn and layer name from a .shp filepath
}
\details{
import_shp
}
\author{
Bruno Silva
}
\keyword{internal}
