% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_clip.R
\name{road_clip}
\alias{road_clip}
\title{Clips a polygon to a buffer}
\usage{
road_clip(kernel_shape, road_shape, buf_width)
}
\arguments{
\item{kernel_shape}{Density kernel shapefile}

\item{road_shape}{Roads shapefile}

\item{buf_width}{Buffer width in maps units}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Clips a density kernel to a buffer around linear features
}
\details{
road_kernel
}
\author{
Bruno Silva
}
\keyword{internal}
