% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_kernel.R
\name{road_kernel}
\alias{road_kernel}
\title{Title}
\usage{
road_kernel(count_path, roads_path, bandw = 500)
}
\arguments{
\item{count_path}{}

\item{roads_path}{}

\item{bandw}{}
}
\value{
string
}
\description{
Title
}
\examples{
not run
}

