% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Interactive visualization of hotspots}
\usage{
run_app()
}
\value{
Nothing
}
\description{
Shiny app to visualize road hotspots using both malo method and a gaussian density kernel
}
\details{
run_app
}
\author{
Bruno Silva
}
