% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_group.R
\name{select_group}
\alias{select_group}
\title{Select rows from dataset}
\usage{
select_group(data, group = "all")
}
\arguments{
\item{data}{The dataframe to subset}

\item{group}{String to be matched in first column. 
Can be a single string or multiple strings. For multiple
strings use: c("string1", "string2"). By default the function
returns exactly the original dataframe.}
}
\value{
A dataframe with the selected rows
}
\description{
Select rows from dataset by values of 
first column
}
\details{
select_group
}
\author{
Bruno Silva
}
\keyword{internal}
