% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_shp.R
\name{import_shp}
\alias{import_shp}
\title{Creates a valid dsn and layer name from a .shp filepath}
\usage{
import_shp(path)
}
\arguments{
\item{path}{The full path of a .shp file}
}
\value{
List with two elements: dsn and layer name
}
\description{
Creates a valid dsn and layer name from a .shp filepath
}
\examples{
import_shp("home/user/data/shapefile.shp")
}
\author{
Bruno Silva
}

