% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines2df.R
\name{lines2df}
\alias{lines2df}
\title{SpatialLines to dataframe}
\usage{
lines2df(spatial_line)
}
\arguments{
\item{spatial_line}{An object class SpatialLinesDataFrame from the package sp.}
}
\value{
SpatialLinesDataFrame
}
\description{
This function extracts the coordinates of the start and end points of each segment of a Spatial Line and returns a dataframe.
}
\details{
lines2df
}
\author{
Duccio Aiazzi
}
\keyword{internal}
