% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_kernel.R
\name{road_kernel}
\alias{road_kernel}
\title{Kernel density estimation on roads}
\usage{
road_kernel (count_path, roads_path, bandw = 500)
}
\arguments{
\item{count_path}{Path for the .csv file with the location of the observations. 
The file must have three headed columns (id, x coordinate and y coordinate) 
with one observation per row}

\item{roads_path}{Path for the .shp file with the roads}

\item{bandw}{Vector of bandwidths for x and y directions. 
A scalar value will be taken to apply to both directions.}
}
\value{
A SpatialPolygonsDataFrame object with the kernel 
density estimation
}
\description{
Kernel density estimation on roads using the function \code{\link[MASS]{kde2d}}
from package MASS.
}
\details{
Observations are snaped to closest roads prior to  
estimation and the resulting kernel is clipped to the road. The kernel 
density estimation is rescaled to [0, 1] and recoded in 
three categories: 1 = 0.25 to 0.50, 2 = 0.50 to 0.75,
3 = 0.75 to 1
}
\examples{
roads_path <- system.file("extdata/roads.shp", package = "roadHotspots")
count_path <- system.file("extdata/count.csv", package = "roadHotspots")
output <- road_kernel(count_path, roads_path) 
}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied 
Statistics with S. Fourth edition. Springer.
}
\author{
Bruno Silva
}
