% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_clip.R, R/road_kernel.R
\name{road_kernel}
\alias{road_clip}
\alias{road_kernel}
\title{Clips a density kernel to a buffer around linear features}
\usage{
road_clip(kernel_shape, road_shape, buf_width)

road_kernel(count_path, roads_path, bandw = 500)
}
\arguments{
\item{kernel_shape}{Density kernel shapefile}

\item{road_shape}{Roads shapefile}

\item{buf_width}{Buffer width in maps units}

\item{count_path}{Path for the .csv file with the location of the observations}

\item{roads_path}{Path for the .shp file with the roads}

\item{bandw}{Vector of bandwidths for x and y directions. A scalar value will be taken to apply to both directions.}
}
\value{
Clipped density kernel to the roads

string
}
\description{
A Private Function

Title
}
\author{
Bruno Silva

Bruno Silva
}

