% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_malo.R
\name{road_malo}
\alias{road_malo}
\title{Road hotspots using Malo method}
\usage{
road_malo(count_path, roads_path, thresh = 0.95)
}
\arguments{
\item{count_path}{Path for the .csv file with the location of the observations}

\item{roads_path}{Path for the .shp file with the roads}

\item{thresh}{Probability threshold hotspot}
}
\value{
a SpatialLinesDataFrame object with the hotspot road segments
}
\description{
Title
}
\author{
Bruno Silva
}
