% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Shiny app to visualize road hotspots}
\usage{
run_app()
}
\value{
Nothing
}
\description{
Shiny app to visualize road hotspots using Malo's method 
or a gaussian kernel density estimation
}
\details{
The app provides a GUI for the functions provided in this package
}
\author{
Bruno Silva
}
