% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lines.R
\name{split_lines}
\alias{split_lines}
\title{Split lines into same length segments}
\usage{
split_lines(spatial_line, split_length = 500)
}
\arguments{
\item{spatial_line}{An "sp" object of class SpatialLines or SpatialLinesDataFrame.}

\item{split_length}{The length of the segments to split the lines into, in units of the SpatialLines. Default 500.}
}
\value{
SpatialLinesDataFrame
}
\description{
This function splits each line of an "sp" object of class SpatialLines into segments of a given length.
}
\details{
split_lines
}
\references{
http://math.stackexchange.com/questions/175896/finding-a-point-along-a-line-a-certain-distance-away-from-another-point?newreg=468f66d7274f449b8ecf3fa4e63f41fe

http://tutorial.math.lamar.edu/Classes/CalcII/Vectors_Basics.aspx
}
\author{
Duccio Aiazzi
}
\keyword{internal}
